Replace "RadioMediumObserver" with "RadioTransmissionObserver"
You fail to build cooja-radiologger-headless under the master branch
of contiki because of the refactoring of RadioMedium commited to
contiki/master on Aug 31, 2015. The short hash of the commit is
c1a275f.

Two methods on which cooja-radiologger-headless depends,
"addRadioMediumObserver" and "deleteRadioMediumObserver, are
obsolete. They was changed to "addRadioTransmissionObserver" and
"deleteRadioTransmissionObserver" by the commit.

In order to follow the changes, the necessary string replacement was
made.